refactor: split openai adapter into openai/ folder
