+++ apps/backend/src/agent-core/llm-api/openai/index.ts
@@ -1,0 +1,1 @@
+export {streamOpenAI} from './stream.js';
