+++ apps/backend/src/agent-core/llm-api/openai/stream.ts
@@ -1,67 +1,10 @@
 import assert from 'node:assert';
 
-import type {ThinkingLevel} from '@omnicraft/api-schema';
 import type OpenAI from 'openai';
 import OpenAIClient from 'openai';
-import {z} from 'zod';
 
-import type {ToolDefinition} from '../tool/types.js';
-import type {
-  LlmCompletionOptions,
-  LlmEventStream,
-  LlmMessage,
-} from './types.js';
-
-type SdkMessageParam = OpenAI.ChatCompletionMessageParam;
-
-/** Converts our unified LlmMessage to the OpenAI SDK message format. */
-function toSdkMessage(message: LlmMessage): SdkMessageParam {
-  switch (message.role) {
-    case 'user':
-      return {role: 'user', content: message.content};
-    case 'assistant': {
-      const toolCalls =
-        message.toolCalls.length > 0
-          ? message.toolCalls.map((tc) => ({
-              id: tc.callId,
-              type: 'function' as const,
-              function: {name: tc.toolName, arguments: tc.arguments},
-            }))
-          : undefined;
-      return {
-        role: 'assistant',
-        content: message.content,
-        tool_calls: toolCalls,
-      };
-    }
-    case 'tool':
-      return {
-        role: 'tool',
-        tool_call_id: message.callId,
-        content: message.content,
-      };
-  }
-}
-
-/** Converts a ToolDefinition to the OpenAI tool format. */
-function toOpenAITool(tool: ToolDefinition): OpenAI.ChatCompletionTool {
-  return {
-    type: 'function',
-    function: {
-      name: tool.name,
-      description: tool.description,
-      parameters: z.toJSONSchema(tool.parameters),
-    },
-  };
-}
-
-/** Maps a ThinkingLevel to the OpenAI reasoning_effort value. */
-function toReasoningEffort(
-  level: ThinkingLevel,
-): 'low' | 'medium' | 'high' | undefined {
-  if (level === 'none') return undefined;
-  return level;
-}
+import type {LlmCompletionOptions, LlmEventStream} from '../types.js';
+import {toOpenAITool, toReasoningEffort, toSdkMessage} from './helpers.js';
 
 /** Streams LLM events from an OpenAI-compatible API. */
 export async function* streamOpenAI(
@@ -96,7 +39,7 @@ export async function* streamOpenAI(
   );
 
   // OpenAI streams tool calls incrementally per choice delta.
-  // Track index → callId so we can emit tool-call-end with the correct callId.
+  // Track index -> callId so we can emit tool-call-end with the correct callId.
   const callIdsByIndex = new Map<number, string>();
   let isFirstChunk = true;
   let lastFinishReason = '';
